+++ packages/react-learn-card/src/components/CertificateDisplayCard/CertificateFrontFace.tsx
@@ -22,6 +22,7 @@ import VerifierStateBadgeAndText, {
     VerifierState,
     VERIFIER_STATES,
 } from './VerifierStateBadgeAndText';
+import { KnownDIDRegistryType } from '../../types';
 
 type CertificateFrontFaceProps = {
     isFront?: boolean;
@@ -29,7 +30,7 @@ type CertificateFrontFaceProps = {
     categoryType?: LCCategoryEnum;
     issuerOverride?: Profile;
     issueeOverride?: Profile;
-    trustedAppRegistry?: any[];
+    knownDIDRegistry?: KnownDIDRegistryType;
     subjectImageComponent?: React.ReactNode;
     issuerImageComponent?: React.ReactNode;
     customBodyCardComponent?: React.ReactNode;
@@ -45,7 +46,7 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
     categoryType,
     issuerOverride,
     issueeOverride,
-    trustedAppRegistry,
+    knownDIDRegistry,
     subjectImageComponent,
     issuerImageComponent,
     customBodyCardComponent,
@@ -78,20 +79,20 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
     let borderColor = `border-${credentialLightColor}`;
 
     if (categoryType === LCCategoryEnum.accommodations) {
-        textLightColor = 'text-amber-500';
-        textDarkColor = 'text-amber-700';
-        borderColor = 'border-amber-500';
+        textLightColor = 'text-violet-500';
+        textDarkColor = 'text-violet-700';
+        borderColor = 'border-violet-500';
     } else if (categoryType === LCCategoryEnum.accomplishments) {
-        textLightColor = 'text-lime-500';
-        textDarkColor = 'text-lime-700';
-        borderColor = 'border-lime-500';
+        textLightColor = 'text-yellow-500';
+        textDarkColor = 'text-yellow-700';
+        borderColor = 'border-yellow-500';
     } else if (categoryType === LCCategoryEnum.learningHistory) {
-        categoryTitle = 'Course';
+        categoryTitle = 'Study';
     } else if (categoryType === LCCategoryEnum.workHistory) {
         categoryTitle = 'Experiences';
-        textLightColor = 'text-blue-500';
-        textDarkColor = 'text-blue-700';
-        borderColor = 'border-blue-500';
+        textLightColor = 'text-cyan-500';
+        textDarkColor = 'text-cyan-700';
+        borderColor = 'border-cyan-500';
     }
 
     const issuerName = getNameFromProfile(issuer ?? '');
@@ -108,14 +109,12 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
         // the did:example:123 condition is so that we don't show this status from the Manage Boosts tab
         verifierState = VERIFIER_STATES.selfVerified;
     } else {
-        const appRegistryEntry = trustedAppRegistry?.find(
-            registryEntry => registryEntry.did === issuerDid
-        );
-
-        if (appRegistryEntry) {
-            verifierState = appRegistryEntry.isTrusted
-                ? VERIFIER_STATES.trustedVerifier
-                : VERIFIER_STATES.untrustedVerifier;
+        if (knownDIDRegistry?.source === 'trusted') {
+            verifierState = VERIFIER_STATES.trustedVerifier;
+        } else if (knownDIDRegistry?.source === 'untrusted') {
+            verifierState = VERIFIER_STATES.untrustedVerifier;
+        } else if (knownDIDRegistry?.source === 'unknown') {
+            verifierState = VERIFIER_STATES.unknownVerifier;
         } else {
             verifierState = VERIFIER_STATES.unknownVerifier;
         }
@@ -128,7 +127,7 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
         <section
             role="button"
             onClick={() => handleViewBackFace?.()}
-            className="relative p-[13px] mt-[55px] bg-white border-[5px] rounded-[30px] border-soid border-grayscale-200 md:max-w-[300px] sm:max-w-[500px]"
+            className="relative p-[13px] mt-[55px] bg-white border-[5px] rounded-[30px] border-soid border-grayscale-200"
         >
             <div className="w-[calc(100%-26px)] absolute top-[-52px]">
                 <CertificateImageDisplay
@@ -213,9 +212,7 @@ export const CertificateFrontFace: React.FC<CertificateFrontFaceProps> = ({
 
                     <VerifierStateBadgeAndText verifierState={verifierState} />
                 </div>
-                <div
-                    className={`${textLightColor} uppercase text-[14px] font-notoSans font-[600]`}
-                >
+                <div className={`${textLightColor} uppercase text-[14px] font-notoSans font-[600]`}>
                     {categoryTitle}
                 </div>
             </div>