+++ services/learn-card-network/brain-service/src/app.ts
@@ -1,19 +1,31 @@
 import { t } from '@routes';
-import { boostsRouter } from '@routes/boosts';
-import { claimHooksRouter } from '@routes/claim-hooks';
-import { profilesRouter } from '@routes/profiles';
-import { profileManagersRouter } from '@routes/profile-manager';
-import { credentialsRouter } from '@routes/credentials';
-import { presentationsRouter } from '@routes/presentations';
-import { storageRouter } from '@routes/storage';
-import { utilitiesRouter } from '@routes/utilities';
-import { contractsRouter } from '@routes/contracts';
-import { didMetadataRouter } from '@routes/did-metadata';
-import { authGrantsRouter } from '@routes/auth-grants';
+import { boostsRouter, BoostsRouter } from '@routes/boosts';
+import { claimHooksRouter, ClaimHooksRouter } from '@routes/claim-hooks';
+import { profilesRouter, ProfilesRouter } from '@routes/profiles';
+import { profileManagersRouter, ProfileManagersRouter } from '@routes/profile-manager';
+import { credentialsRouter, CredentialsRouter } from '@routes/credentials';
+import { presentationsRouter, PresentationsRouter } from '@routes/presentations';
+import { storageRouter, StorageRouter } from '@routes/storage';
+import { utilitiesRouter, UtilitiesRouter } from '@routes/utilities';
+import { contractsRouter, ContractsRouter } from '@routes/contracts';
+import { didMetadataRouter, DidMetadataRouter } from '@routes/did-metadata';
+import { authGrantsRouter, AuthGrantsRouter } from '@routes/auth-grants';
 
 export { createContext } from '@routes';
 
-export const appRouter = t.router({
+export const appRouter = t.router<{
+    boost: BoostsRouter;
+    claimHook: ClaimHooksRouter;
+    profile: ProfilesRouter;
+    profileManager: ProfileManagersRouter;
+    credential: CredentialsRouter;
+    presentation: PresentationsRouter;
+    storage: StorageRouter;
+    utilities: UtilitiesRouter;
+    contracts: ContractsRouter;
+    didMetadata: DidMetadataRouter;
+    authGrants: AuthGrantsRouter;
+}>({
     boost: boostsRouter,
     claimHook: claimHooksRouter,
     profile: profilesRouter,